+++ lib/omniauth-azure-devops.rb
